% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubbleplot}
\alias{bubbleplot}
\title{Bubble plot}
\usage{
bubbleplot(CountDiagnosis, topn)
}
\arguments{
\item{CountDiagnosis}{Dataframe; one case in \code{Bubble_data} output from \code{\link{main}}.}

\item{topn}{Integer; if \code{topn} != NULL, the y-axis of the bubble plot only contains the most frequent \code{topn} diseases.}
}
\value{
A ggplot object.
}
\description{
Create a bubble plot for count data
}
