% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labplot.R
\name{labplot}
\alias{labplot}
\title{Density plot}
\usage{
labplot(
  PatientObersvations_pro,
  loinc_mapping,
  windows.size,
  windows.min,
  windows.max
)
}
\arguments{
\item{PatientObersvations_pro}{Dataframe; \code{PatientObersvations_pro} output from \code{\link{main}}.}

\item{loinc_mapping}{Dataframe; connecting loinc codes to detailed description.}
}
\value{
A ggplot object.
}
\description{
Create density plot for lab data
}
