% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loinc_mapping}
\alias{loinc_mapping}
\title{Loinc Lab mapping data.}
\format{
An object of class \code{data.frame} with 18 rows and 5 columns.
}
\usage{
loinc_mapping
}
\description{
A dataset containing the Loinc lab code and description.
}
\keyword{datasets}
