% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Creat bubble plots and density plots}
\usage{
main(
  PatientSummary,
  PatientObservations,
  PheCodes,
  loinc_mapping,
  digits = NULL,
  windows.size.bubble = 30,
  windows.size.density = 30,
  windows.min = 0,
  windows.max = 120,
  topn = NULL
)
}
\arguments{
\item{PatientSummary}{Dataframe; provided in the 4CE 2.1 data.}

\item{PatientObservations}{Dataframe; provided in the 4CE 2.1 data.}

\item{PheCodes}{Dataframe; a mapping file to roll up ICD codes to the Phecode level.}

\item{loinc_mapping}{Dataframe; connecting loinc codes to detailed description.}

\item{digits}{Integer; the digit of ICD code. If default = NULL, set to
be the largest digit of numbers in the \code{concept_code} column in \code{PatientObservations}.}

\item{windows.size.bubble}{Integer; the size of each window in the bubble plot, default=30.}

\item{windows.size.density}{Integer; the size of each window in the density plot, default=30.}

\item{windows.min}{Integer; the minimum time point in the bubble plot, default = 0.}

\item{windows.max}{Integer; the maximum time point in the bubble plot, default = 120.}

\item{topn}{Integer; number of the most frequently diagnosed diseases to display in the bubble plot (default=NULL).}
}
\value{
A list with the following components:

\tabular{ll}{
\code{data} \tab Processed \code{PatientObservations} with rollup information and input data for bubbleplot. \cr
\code{bubble} \tab Bubble plots for ICD and LAB count data with 4 cases.  \cr
\code{density} \tab Density plots for continuous LAB data. \cr
}
}
\description{
Generate data frames that count the number of patients diagnosed with different diseases
under different cases within time windows. Create bubble plots and density plots for rollup ICD data and Lab data.
}
