% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientlevelplot.R
\name{patientlevelplot}
\alias{patientlevelplot}
\title{Patient-level plot}
\usage{
patientlevelplot(PatientObersvations_pro, patient_num_id)
}
\arguments{
\item{PatientObersvations_pro}{Dataframe; \code{PatientObersvations_pro} output from \code{\link{main}}.}

\item{patient_num_id}{Integer; patient id.}
}
\value{
A ggplot object.
}
\description{
Create patient-level disease count plot over time
}
