% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PatientObservations}
\alias{PatientObservations}
\title{Patient Observations}
\format{
An object of class \code{data.frame} with 30371 rows and 6 columns.

An object of class \code{data.frame} with 30371 rows and 6 columns.

An object of class \code{data.frame} with 30371 rows and 6 columns.
}
\usage{
PatientObservations

PatientObservations

PatientObservations
}
\description{
Simulated PatientObservations file from 4CE v2.1 data.

Simulated PatientClinicalCourse file from 4CE v2.1 data.

Simulated PatientSummary file from 4CE v2.1 data.
}
\keyword{datasets}
