% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PheCodes}
\alias{PheCodes}
\title{ICD mapping data.}
\format{
An object of class \code{data.frame} with 163995 rows and 6 columns.
}
\usage{
PheCodes
}
\description{
A dataset containing the ICD code and Phecode information.
}
\keyword{datasets}
